rename voicecomponent
